SCRUM-17: I implemented the NotFound page in this branch, but I encountered an issue with the SiteHeader. Therefore, I added some comments.
