+++ src/routes/index.tsx
@@ -1,22 +1,10 @@
+import NotFound from "@/components/NotFound"
 import SiteHeader from "@/components/SiteHeader"
 import LoginPage from "@/pages/LoginPage"
 import StudentPage from "@/pages/StudentPage"
 import SubjectPage from "@/pages/SubjectPage"
 import React from "react"
-import { Route, BrowserRouter, Routes, Link } from "react-router-dom"
-
-const NotFound = () => {
-  return (
-      <div className="flex flex-col items-center justify-center min-h-screen bg-gray-100 text-center p-6">
-          <h2 className="text-4xl font-bold text-gray-800 mb-4">404 - Not Found</h2>
-          <p className="text-lg text-gray-600 mb-6">Sorry, we couldn’t find the page you were looking for.</p>
-          <Link to="/login" className="text-blue-500"> 
-          {/* The /login route is temporary while we don't have a Home page. */}
-            Return Home
-          </Link>
-    </div>
-  );
-};
+import { Route, BrowserRouter, Routes } from "react-router-dom"
 
 const AppRoutes = () => {
   return (
